Fail gracefully when no region is configured
If we can't create a client for server side completion,
we should not propogate an exception.  Instead we should
return no server side completion values.

In the future, it would be nice to have some sort of
notification area in the shell where we could let the
user know that server side completion won't work because
they don't have a region configured.

Fixes #84.